Set a better width for .nav_list for alignment with .header_content
This thing of the alignment has been a big issue, maybe I need to
arrange better the elements (the nav in a container with the header
content? maybe it's all about the header).
I set a 800px width through flex-basis and a min-width of 520px for it
to shrink. It also has side margins so when the page is at its smallest,
the nav is perfectly aligned. The issue is that, when the page is being
resized, the alignment is very inconsistent. It's like the .nav_list
and the .header_content grow at a different ratio thoguh are perfectly
aligned at the min and max windows.