Change font's url and fix quote's style
The original URL made it impossible to use Robot in a thin weight.
I styled the quote's new div: its width fits its content, so with
right text alignment, the cite goes to the right. I also had to make the
cite element to be inline-block so I can set a little more top margin
directly to it, and not to use another declaration for the p (its
parent).