Add a script that calculates nav's margins for alignment
The script sets a left margin calculating the distance between the
header title and the page's edge, and does the same for the right margin
but taking the header image and other thigs into account. It's necessary
to substract 10px to the left and 20px to the right for exact alignment.

I also had to change the nav style because it's not necessary anymore
it being a flexbox. Prior .nav_list margins are also not necessary.